Actualización modulo empleados.

+++ app/Http/Controllers/EventoController.php
@@ -9,20 +9,25 @@ use App\Models\FaseTarea;
 use App\Models\Obra;
 use App\Http\Requests\CronogramaRequestSave;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use RealRashid\SweetAlert\Facades\Alert;
 
 class EventoController extends Controller
 {
     public function index($obra)
     {
+        $this->authorize('AccessActividad', Actividad::class);
         $estadoA = EstadoActividad::get()->sortBy('id');
         $faseT = FaseTarea::get()->sortBy('id');
         $obrass = Obra::all();
+        $obra = Obra::find($obra);
+        $actividadBread = $obra->Actividades()->where('Actividads.isActive','=','Active')->get()->first();
         return view('funcionalidades.calendar.index',[
             'estadoA' => $estadoA,
             'faseT' => $faseT,
             'obrasdisp' => $obrass,
-            'idobr' => $obra
+            'idobr' => $obra,
+            'actividadBread' => $actividadBread
         ]);
     }
 
@@ -41,6 +46,7 @@ class EventoController extends Controller
 
     public function show(Evento $evento)
     {
+        $this->authorize('ShowActividad', Actividad::class);
         return response()->json($evento);
     }
 
@@ -49,6 +55,7 @@ class EventoController extends Controller
 
     public function store($obra, CronogramaRequestSave $request)
     {
+        $this->authorize('CreateActividad', Actividad::class);
         $actividad = Actividad::updateOrCreate(['id' => $request->id],[
             'title' => $request->title,
             'DescripcionActividad' => $request->description,
@@ -67,6 +74,7 @@ class EventoController extends Controller
 
     public function edit($obra, Actividad $evento)
     {
+        $this->authorize('UpdateActividad', Actividad::class);
         $evento->start = date('Y-m-d\TH:i', strtotime($evento->start));
         $evento->end = date('Y-m-d\TH:i', strtotime($evento->end));
         if ($evento->obra_id == $obra) {
@@ -76,6 +84,7 @@ class EventoController extends Controller
 
     public function update(Request $request, Actividad $evento)
     {
+        $this->authorize('UpdateActividad', Actividad::class);
         $evento->update($request->validate([
             'title' => 'required',
             'DescripcionActividad' => 'required',
@@ -90,6 +99,7 @@ class EventoController extends Controller
 
     public function destroy(Evento $evento)
     {
+        $this->authorize('DeleteActividad', Actividad::class);
         $evento->delete();
         return response()->json($evento);
     }
